implement request correction modal contents
